Wrapped up database upgrade error handling and related. Plus we have a new version of TBX. High level list of changes:
* Lots of error handling and reporting fixes with DB upgrades.
* Did the code sweep for changing 'Tag' to 'Category'.
* Did the code sweep for changing ID columns from 'int' to 'long'.
* Taking advantage of TBX 3.0.0.7 features:
  - New Debug and Error logging.
  - Support for log limiting.
  - Using TBX's Exception stack from handling.
* Upgraded SQLite from 1.0.82.0 to 1.0.86.0.

-----------------------------------------------------------------------
TESTING
- 
-----------------------------------------------------------------------
DETAILS

- app.config (edit)
  * Nothing, really.

- Classes/Database.cs (edit)
  * All DBI id values changed from 'int' to 'long'.
  * Changed some logging from Info to Debug, now that I have a Debug.
  * Fixed a bug where I used a Query before checking if Query was null

- Classes/Datafile.Downgrade.cs (edit)
  * Finished (but didn't test) SaveAs21.

- Classes/Datafile.Upgrade.cs (edit)
  * Improved error handling and reporting. (Though "improved" is being
    a bit kind. None of my database inserts had *any* error handling
    at all. Just making checks there was the first and most important
    step to getting this usable.
  * Added local CreateNewTable to call CreateTable and set the step.
  * Added SetStep for just that.
  * Fixed a bug where the initial row in old 'projects' tables had
    a null timestamp_c.
  * Topped off migration of everything else: especially splitting up
    the old grid_views into FilterOptions and GridOptions.

- Classes/Diary.cs (edit)
- Classes/Entry.cs (edit)
- Classes/FilterOptions.cs (edit)
- Classes/OtherAttributes.cs (edit)
  * Converted Tag to Category.


- Classes/Item.cs (edit)
  * Converted 'int' to 'long'

- Classes/Options.cs (edit)
  * Added a fourth default option for LastReportView.
  * Fixed the bug where the constructor would always fail loading data.

- Classes/Timekeeper.cs (edit)
  * Moved Abbreviate() to Technitivity.Toolbox.
  * Changed OpenDatabase to match the DBI constructor changes (logLevel)
  * Moved the guts of Exception to TBX.
  * Rearranged and reorganized the file a bit.
  * Added a singular GetLog()

- Forms/fGrid.cs (edit)
  * Converted 'int' to 'long'

- Forms/Filtering.cs (edit)
- Forms/Filtering.Designer.cs (edit)
- Forms/Report.cs (edit)
  * Converted Tag to Category.

- Forms/fMain.Action.cs (edit)
  * Some temporary / WIP changes to opening the database.
  * Converted 'int' to 'long'
  * Better error handling in ACtion_ConvertPriorVersion.
  * Converted Tag to Category.
  * Converted Timekeeper.Abbreviate to Common.Abbreviate.

- Properties/Settings.Designer.cs (edit)
  * Runtime Version update (?)

- Properties/Settings.settings (edit)
  * Lost Profile Name "(Default)" (?)

- Resources/SQL/3.0.0.0/Activity.Create.sql (edit)
  * Fixed an index bug (ActivityName became just Name)

- Resources/SQL/3.0.0.0/Project.Create.sql (edit)
  * Fixed an index bug (ProjectName became just Name)

- timekeeper.csproj (edit)
  * Changed assembly name from 'timekeeper' to 'Timekeeper'.
  * This, along with the changes in TBX fixes the bug with the 
    lowercase 'timekeeper' being shown in Common MessageBox calls.
  * Under Publish, changed version from 2.3.0.3 to 3.0.0.0
  * Upgraded SQLite from 1.0.82.0 to 1.0.86.0.
  * Upgraded Technitivity.Toolbox from 3.0.0.6 to 3.0.0.7.
  * Removed Datafile.Create.cs.

- timekeeper.csproj.user (edit)
  * Removed startup command line argument (for now: trying to test 
    with no prior file loaded).

-----------------------------------------------------------------------
Project...: Toolbox
Components: Library
Target....: 3.0.0.7

Tickets Addressed:
  - #1271 - Move SQL to string resources.

Tickets Closed:
  - n/a

[git-p4: depot-paths = "//depot/timekeeper/": change = 271]

+++ timekeeper/Classes/Item.cs
@@ -237,7 +237,7 @@ namespace Timekeeper
 
         //---------------------------------------------------------------------
 
-        public int Delete()
+        public long Delete()
         {
             this.IsDeleted = true;
             return SetStatus("Deleted", this.IsDeleted);
@@ -296,7 +296,7 @@ namespace Timekeeper
 
         //---------------------------------------------------------------------
 
-        public int Hide()
+        public long Hide()
         {
             this.IsHidden = true;
             return SetStatus("Hidden", this.IsHidden);
@@ -343,7 +343,7 @@ namespace Timekeeper
             Row Row = new Row();
             Row["Description"] = newDescription;
             Row["ModifyTime"] = Common.Now();
-            int Count = Data.Update(this.TableName, Row, this.IdColumnName, this.ItemId);
+            long Count = Data.Update(this.TableName, Row, this.IdColumnName, this.ItemId);
 
             if (Count == 1) {
                 this.Description = newDescription;
@@ -381,7 +381,7 @@ namespace Timekeeper
             Row Row = new Row();
             Row["Name"] = newName;
             Row["ModifyTime"] = Common.Now();
-            int Count = Data.Update(this.TableName, Row, this.IdColumnName, this.ItemId);
+            long Count = Data.Update(this.TableName, Row, this.IdColumnName, this.ItemId);
 
             if (Count == 1) {
                 this.Name = newName;
@@ -398,7 +398,7 @@ namespace Timekeeper
             Row Row = new Row();
             Row["ParentId"] = itemId;
             Row["ModifyTime"] = Common.Now();
-            int Count = Data.Update(this.TableName, Row, this.IdColumnName, this.ItemId);
+            long Count = Data.Update(this.TableName, Row, this.IdColumnName, this.ItemId);
 
             return Count == 1 ? 1 : 0;
         }
@@ -439,7 +439,7 @@ namespace Timekeeper
 
         //---------------------------------------------------------------------
 
-        public int Unhide()
+        public long Unhide()
         {
             this.IsHidden = false;
             return SetStatus("Hidden", this.IsHidden);
@@ -494,7 +494,7 @@ namespace Timekeeper
 
         //---------------------------------------------------------------------
 
-        private int SetStatus(string status, bool isSet)
+        private long SetStatus(string status, bool isSet)
         {
             string BooleanColumnName = "Is" + status;
             string TimeColumnName = status + "Time";
